requirements.txt
asgiref==3.4.1
astroid==2.8.4
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Django==3.2.7
django-extensions==3.1.5
django-otp==1.1.3
django-rest-auth==0.9.5
djangorestframework==3.12.4
gunicorn==20.1.0
h11==0.12.0
idna==3.3
isort==5.9.3
lazy-object-proxy==1.6.0
logilab-common==1.8.1
mccabe==0.6.1
mypy-extensions==0.4.3
platformdirs==2.4.0
psycopg2-binary==2.9.2
pylint==2.0.0
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
qrcode==7.3.1
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
websockets==10.1
wrapt==1.13.3
